tech adjustments 5: reduce early tech techs to adjust pacing anyways etc.
150->140
225->200
300->275
425->400
575->525
750->700
900->875
1200->1150

lowering these values a bit (which were too high btw) allows for a better scaling and pacing, reaching mid/late classical sooner, having iron working sooner which is nice.

trying not to overdo it as difficulties and maps change stuff and i wouldnt want pace to be too fast on the opposite, so for now set as such as i go into the player wonder question in more detail anyways etc.