replace TECH_REFRIGERATION with TECH_ENVIRONMENTALISM, fix the production chooser bar auto-scrolling to the bottom rows when clicking on one of the lower rows,  and other changes - replace TECH_REFRIGERATION with TECH_ENVIRONMENTALISM : Reuses civ4's CIVIC_ENVIRONMENTALISM's button and TECH_REFRIGERATION's quote + a modern later tech so industrial goes faster and to thematically cover this part of history as well (plus modern was quite most importantly while industrial was too slow anyways etc.). Move the forest preserve's gold effect from TECH_BIOENGINEERING to TECH_ENVIRONMENTALISM, as well as the bonus obsoletes from TECH_MASS_MEDIA and TECH_COMPOSITES. Gets iHealth 1. Also, it uses a real wikipedia based entry, not the very bad and poorly framed civ4 ones. - TECH_MARINE_TECHNOLOGY: change prereqs from TECH_REFRIGERATION AND TECH_ELECTRICITY to TECH_ELECTRICITY - TECH_DEPOPULATION: remove the TECH_REFRIGERATION prereq - BUILDINGCLASS_LABORATORY: no longer requires TECH_REFRIGERATION (now only TECH_THEORY_OF_EVOLUTION) - BUILDINGCLASS_SUPERMARKET: no longer require TECH_REFRIGERATION (now only TECH_CORPORATION) - minor tech prereq event adjustments - TECH_COMPOSITES: now also requires TECH_ENVIRONMENTALISM (with TECH_INDUSTRIALISM). As a nice side effect, this helps gate modern tanks rushes that would otherwise be too soon after the first tanks are discovered at TECH_INDUSTRIALISM anyways etc. - TECH_DEPOPULATION: increase iHealth and iHappiness from -8 to -9 - TECH_TRANSHUMANISM: increase iHealth from 1 to 2 - TECH_TERRAFORMING: increase iHappiness from 2 to 3 - update 48 civs DLL - update docs
Update 2025-12-18 (1 hour after release):

- minor hotfix the requiring TECH_CHIVALRY instead of TECH_MATHEMATICS. Version

If you downloaded the AdvCiv-SAS 5210 first released version (2025-12-18 1 hour ago) before, just redownload from same link: it now has the updated file. Since this is a minor change i didn't make a new release for it.

update docs

fix the production chooser bar auto-scrolling to the bottom rows when clicking on one of the lower rows
with the very nice help of chatgpt 5.2 thanks anyways etc.

Note: when at bottom rows and clicking on one of the high rows, we auto-scroll back to top, which is maybe not ideal, or a nice side effect to keep, since we don't need to do much in the lower rows anyway (processes, wonders, etc.), so may as well back scroll to top maybe anyways etc.